implementing visualization for images and csv files containing lists of pkl filepaths

+++ server/mod_binarySegmentation.m
@@ -10,13 +10,13 @@ format compact;
 addpath([pwd, '/Callbacks/'])
 
 %%%% Step 1 Load your data %%%%
-tmp = load([pwd, '/Data/',in_file]);
+tmp = load([pwd, '/Data/', in_file]);
 % data [5xN] 5 channels [Fp1; Fp2; Fpz; F7; F8]
 % t: time number vector in unit [days]
 % Fs: sampling rate
 data = tmp.data;
 t = tmp.t;
-Fs = tmp.Fs;
+Fs = double(tmp.Fs);
 
 %%%% Step 2 Get binary segmentation %%%%    
 [z,a,e,s]=fcn_IdentifyArtifact(data,Fs);
